+++ DependencyInjection/Configuration.php
@@ -28,9 +28,11 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder();
+        $nodeName = 'pcdx_parameter_encryption';
 
-        $rootNode = $treeBuilder->root('pcdx_parameter_encryption');
+        $treeBuilder = new TreeBuilder($nodeName);
+
+        $rootNode = $this->getRootNodeFromTreeBuilder($treeBuilder, $nodeName);
 
         $rootNode
             ->fixXmlConfig('algorithm')
@@ -84,10 +86,13 @@ class Configuration implements ConfigurationInterface
      */
     public function addCryptoNode($type)
     {
-        $builder = new TreeBuilder();
-        $node = $builder->root($type.'ion');
+        $nodeName = $type.'ion';
 
-        $node
+        $treeBuilder = new TreeBuilder($nodeName);
+
+        $rootNode = $this->getRootNodeFromTreeBuilder($treeBuilder, $nodeName);
+
+        $rootNode
             ->info('Configure '.$type.'er.')
             ->isRequired()
             ->fixXmlConfig('argument')
@@ -143,6 +148,23 @@ class Configuration implements ConfigurationInterface
                 ->end()
             ->end();
 
-        return $node;
+        return $rootNode;
+    }
+
+    /**
+     * Get root node from tree builder.
+     *
+     * @deprecated symfony/config 4.2 Simply use TreeBuilder::getRootNode()
+     *
+     * @param TreeBuilder $treeBuilder
+     * @param string      $nodeName
+     *
+     * @return NodeDefinition
+     */
+    private function getRootNodeFromTreeBuilder(TreeBuilder $treeBuilder, string $nodeName): NodeDefinition
+    {
+        return (method_exists($treeBuilder, 'getRootNode')
+            ? $treeBuilder->getRootNode()
+            : $treeBuilder->root($nodeName));
     }
 }